pyproject.toml
[tool.poetry]
name = "django-rest-framework"
version = "0.1.0"
description = ""
authors = ["Gabriel Mango <gabrielhmango@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.0"
django = "^5.0.7"
djangorestframework = "^3.15.2"
markdown = "^3.6"
django-filter = "^24.2"


[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
